Add files via upload
Disability Data set - Looks at if having a disability a factor in overall performance of the student. CSV file
Region Data set - Looks at if the region the student lives in is a factor in overall performance of the student. CSV file
studentInfodw - the code I wrote in python using the pandas to create both csv files. It has other data that counts the number of final_result values (e.g. pass/withdraw/fail/distinction) and groups it by various factors such as age, etc.